requirements.txt
mesop
gunicorn
anthropic
google-generativeai
python-dotenv